Move actions and methods to separate partial controllers
