+++ BTCPayServer/Controllers/UIStoresController.Integrations.cs
@@ -7,7 +7,6 @@ using BTCPayServer.Abstractions.Models;
 using BTCPayServer.Client;
 using BTCPayServer.Client.Models;
 using BTCPayServer.Data;
-using BTCPayServer.Models;
 using BTCPayServer.Models.StoreViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
